guardian-loop/requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.20.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Mechanistic Interpretability
captum>=0.6.0
fancy-einsum>=0.0.3
circuitsvis>=1.43.0

# Martian Integration
martian
requests>=2.31.0

# Adversarial & Rainbow Teaming
nltk>=3.8.0
sentencepiece>=0.1.99

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
ipywidgets>=8.0.0

# Development Tools
jupyter>=1.0.0
streamlit>=1.25.0
gradio>=3.40.0
pytest>=7.4.0
black>=23.0.0
tqdm>=4.65.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
wandb>=0.15.0
einops>=0.6.0 